fix(src/components/markdown.tsx): picture component was being considered as server component
Error reproduction: comment/uncomment 'code' of the components object and save file. A static page
that contains an image in the markdown body will indicate: Error: async/await is not yet supported
in Client Components, only Server Components. This error is often caused by accidentally adding
`'use client'` to a module that was originally written for the server.